[chore] #29 코드 정리
