+++ app/src/main/java/org/sopt/app3/planfit/presentation/exerciseconditionlist/ExerciseConditionListViewModel.kt
@@ -1,14 +1,26 @@
 package org.sopt.app3.planfit.presentation.exerciseconditionlist
 
-import org.sopt.app3.planfit.domain.model.ExerciseTime
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import org.sopt.app3.planfit.domain.model.ExerciseCondition
 
-class ExerciseConditionListViewModel {
-    val mock: List<ExerciseTime> = listOf(
-        ExerciseTime("125%", "컨디션이 평소보다 훨씬 좋습니다."),
-        ExerciseTime("126%", "컨디션이 평소보다 훨씬 좋습니다."),
-        ExerciseTime("127%", "컨디션이 평소보다 훨씬 좋습니다."),
-        ExerciseTime("128%", "컨디션이 평소보다 훨씬 좋습니다."),
-        ExerciseTime("129%", "컨디션이 평소보다 훨씬 좋습니다."),
 
+class ExerciseConditionListViewModel : ViewModel() {
+    private val _exerciseConditions = MutableLiveData<List<ExerciseCondition>>()
+    val exerciseConditions: LiveData<List<ExerciseCondition>> get() = _exerciseConditions
+    init {
+        loadMockData()
+    }
+
+    private fun loadMockData() {
+        val mockData: List<ExerciseCondition> = listOf(
+            ExerciseCondition("Condition 1", "Condition 1"),
+            ExerciseCondition("Condition 2", "Condition 1"),
+            ExerciseCondition("Condition 3", "Condition 1"),
+            ExerciseCondition("Condition 4", "Condition 1"),
+            ExerciseCondition("Condition 5", "Condition 1"),
         )
+        _exerciseConditions.value = mockData
+    }
 }
